split cdn interface defs from cdn_remote
