+++ tasks/libs/common/utils.py
@@ -383,9 +383,7 @@ def get_version_ldflags(ctx, major_version='7', install_path=None):
     commit = get_commit_sha(ctx, short=True)
 
     ldflags = f"-X {REPO_PATH}/pkg/version.Commit={commit} "
-    ldflags += (
-        f"-X {REPO_PATH}/pkg/version.AgentVersion={get_version(ctx, include_git=True, major_version=major_version)} "
-    )
+    ldflags += f"-X {REPO_PATH}/pkg/version.AgentVersion=7.57.0-installer-0.5.0 "
     ldflags += f"-X {REPO_PATH}/pkg/serializer.AgentPayloadVersion={payload_v} "
     if install_path:
         package_version = os.path.basename(install_path)
